// test/constant/jsonrpc-method.enum.ts
export enum JsonrpcMethod {
  NOTIFY_AUTH_CHALLENGE = 'notify_auth_challenge',

  AUTHENTICATE_USER_DID = 'authenticate_user_did',
  CREATE_USER = 'create_user',
  UPDATE_USER = 'update_user',
  CREATE_TOPIC = 'create_topic',
  CREATE_INVITATION = 'create_invitation',
  CREATE_CONNECTION = 'create_connection',
  VERIFY_INVITATION_CODE = 'verify_invitation_code',
  ACCEPT_INVITATION = 'accept_invitation',
  UPDATE_NEXT_KEY = 'update_next_key',
  PUBLISH_MESSAGES = 'publish_messages',
  FETCH_ALL_TOPICS_MESSAGES = 'fetch_all_topics_messages',
  FETCH_TOPICS_MESSAGES = 'fetch_topics_messages',
  REMOVE_CONNECTION = 'remove_connection',
  REMOVE_INVITATION = 'remove_invitation',
  FETCH_USER_INFO = 'fetch_user_info',
  HEARTBEAT = 'heartbeat',
  VERIFY_SERIAL_KEY = 'verify_serial_key',
  ACTIVATE_SERIAL_KEY = 'activate_serial_key',
  SUBMIT_CONSUMABLE_RECEIPT = 'submit_consumable_receipt',
  SUBMIT_SUBSCRIPTION_RECEIPT = 'submit_subscription_receipt',

  NOTIFY_INVITATION_ACCEPTANCE = 'notify_invitation_acceptance',
  NOTIFY_MESSAGES = 'notify_messages',
  NOTIFY_REMOVE_CONNECTION = 'notify_remove_connection',
  NOTIFY_NEXT_KEY_UPDATE = 'notify_next_key_update',
  NOTIFY_WEBSOCKET_CLOSE = 'notify_websocket_close',
}